fake_review_archaeology/requirements.txt
# Fake Review Archaeology - Requirements
# Python 3.9+

# Core Data Science
numpy>=1.24.0
pandas>=1.5.0
scipy>=1.10.0

# Machine Learning
scikit-learn>=1.2.0
xgboost>=1.7.0
lightgbm>=3.3.0

# Deep Learning
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
transformers>=4.30.0
tokenizers>=0.13.0

# NLP
nltk>=3.8.0
textstat>=0.7.0
spacy>=3.5.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0
wordcloud>=1.9.0

# Dashboard
streamlit>=1.24.0
streamlit-echarts>=0.4.0

# Data Processing
pyarrow>=11.0.0
fastparquet>=2023.0.0

# Utilities
tqdm>=4.65.0
python-dotenv>=1.0.0
pyyaml>=6.0
joblib>=1.2.0

# API (Optional)
fastapi>=0.100.0
uvicorn>=0.22.0
pydantic>=2.0.0

# Testing
pytest>=7.3.0
pytest-cov>=4.1.0
hypothesis>=6.75.0

# Development
black>=23.3.0
flake8>=6.0.0
mypy>=1.3.0
pre-commit>=3.3.0

# Jupyter
jupyter>=1.0.0
ipython>=8.13.0
notebook>=6.5.0
